Refactor doQueryJoin and parseJoinArgs to doAddLiquidityQuery and parseAddLiquidityArgs
